test, move dynamic RPC test to e2e test set.
